+++ probe-rs/tests/inlined-functions
